+++ CumulusMX/ExtraSensorSettings.cs
@@ -52,6 +52,7 @@ namespace CumulusMX
 			var ecowittwn34map = new StationSettings.JsonEcowittMappings
 			{
 				primaryTHsensor = cumulus.Gw1000PrimaryTHSensor,
+				primaryIndoorTHsensor = cumulus.Gw1000PrimaryIndoorTHSensor,
 
 				wn34chan1 = cumulus.EcowittMapWN34[1],
 				wn34chan2 = cumulus.EcowittMapWN34[2],
@@ -340,6 +341,7 @@ namespace CumulusMX
 						}
 
 						cumulus.Gw1000PrimaryTHSensor = settings.httpSensors.ecowitt.mappings.primaryTHsensor;
+						cumulus.Gw1000PrimaryIndoorTHSensor = settings.httpSensors.ecowitt.mappings.primaryIndoorTHsensor;
 
 						if (cumulus.EcowittMapWN34[1] != settings.httpSensors.ecowitt.mappings.wn34chan1)
 						{